refactor(reservation): extract step components from page
- Create StepDates, StepGuests, StepConfirm, StepPayment as standalone components
- Update barrel index.ts to re-export all 4 step components
- Replace 4 inline step definitions in ReservationFlowPage with imports
- Clean up 10 unused lucide icons, useState, useCallback, and other dead imports
- Page reduced from 670 to 133 lines
- Zero behavior change

PR3/4 of refactor/reservation-flow chain
