+++ frontend/src/components/reservation/index.ts
@@ -7,3 +7,7 @@
 export { default as PriceBadge } from './price-badge';
 export { default as PriceBreakdownCard } from './price-breakdown-card';
 export { default as StepIndicator } from './step-indicator';
+export { default as StepDates } from './step-dates';
+export { default as StepGuests } from './step-guests';
+export { default as StepConfirm } from './step-confirm';
+export { default as StepPayment } from './step-payment';
