+++ frontend/src/pages/ReservationFlowPage.tsx
@@ -6,32 +6,20 @@
  * @author Nexo Real Development Team
  */
 
-import { useEffect, useState, useMemo, useCallback } from 'react';
+import { useEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router';
 import { useTranslation } from 'react-i18next';
-import {
-  Users,
-  CheckCircle2,
-  ChevronRight,
-  ArrowLeft,
-  MapPin,
-  Compass,
-  Loader2,
-  AlertCircle,
-  Wallet,
-  Shield,
-  Clock,
-  Lock,
-} from 'lucide-react';
+import { MapPin, Compass } from 'lucide-react';
 import { Button } from '../components/ui/button';
-import { useReservationWizard } from '../stores/reservationStore';
-import { computePriceBreakdown, formatPrice } from '../stores/reservationStore';
-import type { WizardStep, PriceBreakdown } from '../stores/reservationStore';
-import { PriceBadge, PriceBreakdownCard, StepIndicator } from '../components/reservation';
-import { paymentService } from '../services/paymentService';
-import { useWalletBalance } from '../stores/walletStore';
-import { cn } from '../lib/utils';
-import { featureFlags } from '../utils/featureFlags';
+import { useReservationWizard, computePriceBreakdown } from '../stores/reservationStore';
+import type { WizardStep } from '../stores/reservationStore';
+import {
+  StepIndicator,
+  StepDates,
+  StepGuests,
+  StepConfirm,
+  StepPayment,
+} from '../components/reservation';
 
 // ============================================
 // Constants / Constantes
@@ -47,525 +35,6 @@ const STEPS: { id: WizardStep; labelKey: string }[] = [
   { id: 'confirm', labelKey: 'reservation.confirmation' },
   { id: 'payment', labelKey: 'reservation.payment' },
 ];
-
-// ============================================
-// Step 1: Dates (property only)
-// ============================================
-
-function StepDates() {
-  const { t } = useTranslation();
-  const { wizardData, updateWizardData, setWizardStep } = useReservationWizard();
-
-  const breakdown = useMemo(() => computePriceBreakdown(wizardData), [wizardData]);
-
-  if (!wizardData || wizardData.type !== 'property') return null;
-
-  const canContinue = wizardData.checkIn && wizardData.checkOut;
-
-  return (
-    <div className="space-y-6">
-      <div>
-        <h2 className="text-xl font-bold text-slate-800 mb-1">{t('reservation.selectDates')}</h2>
-        <p className="text-slate-500 text-sm">
-          {t('reservation.property')}:{' '}
-          <span className="font-medium text-slate-700">{wizardData.property.title}</span>
-        </p>
-      </div>
-
-      {/* Price per night indicator */}
-      {breakdown && (
-        <div className="rounded-lg border border-emerald-200 bg-emerald-50/50 px-4 py-3">
-          <span className="text-sm text-slate-600">
-            {t('reservation.pricePerNight')}:{' '}
-            <span className="font-bold text-emerald-600">
-              {formatPrice(breakdown.pricePerUnit, breakdown.currency)}
-            </span>
-          </span>
-        </div>
-      )}
-
-      <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-        <div>
-          <label className="block text-sm font-medium text-slate-700 mb-1">
-            {t('reservation.checkIn')}
-          </label>
-          <input
-            type="date"
-            value={wizardData.checkIn}
-            min={new Date().toISOString().split('T')[0]}
-            onChange={(e) => updateWizardData({ checkIn: e.target.value })}
-            className="w-full px-3 py-2 rounded-lg border border-slate-200 text-sm focus:outline-none focus:border-emerald-400"
-          />
-        </div>
-        <div>
-          <label className="block text-sm font-medium text-slate-700 mb-1">
-            {t('reservation.checkOut')}
-          </label>
-          <input
-            type="date"
-            value={wizardData.checkOut}
-            min={wizardData.checkIn || new Date().toISOString().split('T')[0]}
-            onChange={(e) => updateWizardData({ checkOut: e.target.value })}
-            className="w-full px-3 py-2 rounded-lg border border-slate-200 text-sm focus:outline-none focus:border-emerald-400"
-          />
-        </div>
-      </div>
-
-      {/* Running total after selecting dates */}
-      {breakdown && breakdown.totalNights > 0 && <PriceBadge breakdown={breakdown} />}
-
-      <Button
-        onClick={() => setWizardStep('guests')}
-        disabled={!canContinue}
-        className="w-full gap-2 py-3"
-      >
-        {t('reservation.continue')}
-        <ChevronRight className="w-5 h-5" />
-      </Button>
-    </div>
-  );
-}
-
-// ============================================
-// Step 2: Guests
-// ============================================
-
-function StepGuests() {
-  const { t } = useTranslation();
-  const {
-    wizardData,
-    updateWizardData,
-    setWizardStep,
-    isCreating,
-    createError,
-    confirmReservation,
-  } = useReservationWizard();
-
-  const breakdown = useMemo(() => computePriceBreakdown(wizardData), [wizardData]);
-
-  if (!wizardData) return null;
-
-  const maxGuests = wizardData.type === 'tour' ? wizardData.availability.availableSpots : 20;
-
-  const handleConfirm = async () => {
-    try {
-      await confirmReservation();
-    } catch {
-      // Error is already captured in createError state by the store.
-      // El store captura el error en createError — no se necesita manejo adicional aquí.
-    }
-  };
-
-  return (
-    <div className="space-y-6">
-      <div>
-        <h2 className="text-xl font-bold text-slate-800 mb-1">{t('reservation.howManyGuests')}</h2>
-        {wizardData.type === 'property' ? (
-          <p className="text-slate-500 text-sm">
-            {t('reservation.property')}:{' '}
-            <span className="font-medium text-slate-700">{wizardData.property.title}</span>
-          </p>
-        ) : (
-          <p className="text-slate-500 text-sm">
-            {t('reservation.tour')}:{' '}
-            <span className="font-medium text-slate-700">{wizardData.tour.title}</span>
-            {' · '}
-            {new Date(wizardData.availability.date).toLocaleDateString('es-AR', {
-              weekday: 'long',
-              day: 'numeric',
-              month: 'long',
-            })}
-          </p>
-        )}
-      </div>
-
-      {/* Guest counter */}
-      <div className="flex items-center gap-4">
-        <label className="text-sm font-medium text-slate-700 flex items-center gap-2">
-          <Users className="w-4 h-4 text-emerald-500" />
-          {t('reservation.numberOfGuests')}
-        </label>
-        <div className="flex items-center gap-3 ml-auto">
-          <Button
-            variant="outline"
-            size="icon"
-            onClick={() => updateWizardData({ guests: Math.max(1, wizardData.guests - 1) })}
-            className="h-11 w-11 rounded-lg text-lg font-medium"
-          >
-            −
-          </Button>
-          <span className="w-10 text-center font-semibold text-lg text-slate-800">
-            {wizardData.guests}
-          </span>
-          <Button
-            variant="outline"
-            size="icon"
-            onClick={() => updateWizardData({ guests: Math.min(maxGuests, wizardData.guests + 1) })}
-            className="h-11 w-11 rounded-lg text-lg font-medium"
-          >
-            +
-          </Button>
-        </div>
-      </div>
-
-      {/* Running price total */}
-      {breakdown && breakdown.totalPrice > 0 && <PriceBadge breakdown={breakdown} />}
-
-      {/* Notes */}
-      <div>
-        <label className="block text-sm font-medium text-slate-700 mb-1">
-          {t('reservation.additionalNotes')}{' '}
-          <span className="text-slate-400 font-normal">({t('reservation.optional')})</span>
-        </label>
-        <textarea
-          value={wizardData.notes}
-          onChange={(e) => updateWizardData({ notes: e.target.value })}
-          placeholder={t('reservation.additionalNotesHint')}
-          rows={3}
-          className="w-full px-3 py-2 rounded-lg border border-slate-200 text-sm focus:outline-none focus:border-emerald-400 resize-none"
-        />
-      </div>
-
-      {/* Error */}
-      {createError && (
-        <div className="flex items-start gap-2 bg-red-50 border border-red-200 rounded-lg p-3 text-sm text-red-600">
-          <AlertCircle className="w-4 h-4 mt-0.5 shrink-0" />
-          {createError}
-        </div>
-      )}
-
-      <div className="flex gap-3">
-        {wizardData.type === 'property' && (
-          <Button
-            variant="outline"
-            onClick={() => setWizardStep('dates')}
-            className="gap-2 px-4 py-3"
-          >
-            <ArrowLeft className="w-4 h-4" />
-            {t('reservation.back')}
-          </Button>
-        )}
-        <Button onClick={handleConfirm} disabled={isCreating} className="flex-1 gap-2 py-3">
-          {isCreating ? (
-            <>
-              <Loader2 className="w-5 h-5 animate-spin" />
-              {t('reservation.confirming')}
-            </>
-          ) : (
-            <>
-              {t('reservation.confirmReservation')}
-              <ChevronRight className="w-5 h-5" />
-            </>
-          )}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// ============================================
-// Step 3: Confirmation (success + proceed to payment)
-// ============================================
-
-interface StepConfirmProps {
-  onGoToPayment: () => void;
-  onGoToReservations: () => void;
-  breakdown: PriceBreakdown | null;
-}
-
-function StepConfirm({ onGoToPayment, onGoToReservations, breakdown }: StepConfirmProps) {
-  const { t } = useTranslation();
-  const { createdReservation, wizardData } = useReservationWizard();
-
-  return (
-    <div className="space-y-6 py-2">
-      {/* Success header */}
-      <div className="text-center">
-        <div className="w-20 h-20 bg-emerald-100 rounded-full flex items-center justify-center mx-auto mb-4">
-          <CheckCircle2 className="w-10 h-10 text-emerald-500" />
-        </div>
-        <h2 className="text-2xl font-bold text-slate-800 mb-2">
-          {t('reservation.reservationConfirmed')}
-        </h2>
-        <p className="text-slate-500 text-sm">{t('reservation.reservationConfirmedDesc')}</p>
-      </div>
-
-      {/* Reservation details */}
-      {createdReservation && (
-        <div className="bg-slate-50 border border-slate-200 rounded-xl p-4 text-left space-y-2">
-          <p className="text-sm text-slate-500">
-            <span className="font-medium text-slate-700">{t('reservation.reservationId')}:</span>{' '}
-            <span className="font-mono text-xs">{createdReservation.id}</span>
-          </p>
-          <p className="text-sm text-slate-500">
-            <span className="font-medium text-slate-700">{t('reservation.status')}:</span>{' '}
-            <span className="capitalize">{createdReservation.status}</span>
-          </p>
-          {wizardData?.type === 'property' && (
-            <p className="text-sm text-slate-500">
-              <span className="font-medium text-slate-700">{t('reservation.property')}:</span>{' '}
-              {wizardData.property.title}
-            </p>
-          )}
-          {wizardData?.type === 'tour' && (
-            <p className="text-sm text-slate-500">
-              <span className="font-medium text-slate-700">{t('reservation.tour')}:</span>{' '}
-              {wizardData.tour.title}
-            </p>
-          )}
-        </div>
-      )}
-
-      {/* Price breakdown card */}
-      {breakdown && breakdown.totalPrice > 0 && <PriceBreakdownCard breakdown={breakdown} />}
-
-      {/* Action buttons */}
-      <div className="flex flex-col gap-3">
-        <Button onClick={onGoToPayment} className="w-full gap-2 py-3">
-          <Lock className="mr-2 h-4 w-4" />
-          {t('reservation.selectPayment')}
-          <ChevronRight className="w-5 h-5" />
-        </Button>
-        <Button variant="ghost" onClick={onGoToReservations} className="w-full py-3 text-sm">
-          {t('reservation.payLater')}
-          <span className="ml-2 text-slate-400">— {t('reservation.payLaterHint')}</span>
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// ============================================
-// Step 4: Payment method selector
-// ============================================
-
-interface StepPaymentProps {
-  onDone: () => void;
-  onGoToReservations: () => void;
-  breakdown: PriceBreakdown | null;
-}
-
-function StepPayment({ onDone, onGoToReservations, breakdown }: StepPaymentProps) {
-  const { t } = useTranslation();
-  const {
-    createdReservation,
-    isProcessingPayment,
-    paymentError,
-    setPaymentProcessing,
-    setPaymentError,
-  } = useReservationWizard();
-  const { balance } = useWalletBalance();
-  const [processingMethod, setProcessingMethod] = useState<string | null>(null);
-
-  const walletBalance = balance?.balance ?? 0;
-  const walletCurrency = balance?.currency ?? 'USD';
-  const totalPrice = breakdown?.totalPrice ?? 0;
-  const currency = breakdown?.currency ?? 'USD';
-  const hasEnoughBalance = walletBalance >= totalPrice && totalPrice > 0;
-
-  const handlePayPal = useCallback(async () => {
-    if (!createdReservation || !breakdown) return;
-    setProcessingMethod('paypal');
-    setPaymentProcessing(true);
-    setPaymentError(null);
-
-    try {
-      const result = await paymentService.createPayPalOrder({
-        amount: breakdown.totalPrice,
-        currency: breakdown.currency,
-        description: `Nexo Real - Reservation ${createdReservation.id}`,
-        orderId: createdReservation.id,
-      });
-
-      if (result.data.approvalUrl) {
-        window.location.href = result.data.approvalUrl;
-      }
-    } catch {
-      setPaymentError(t('reservation.paymentError'));
-    } finally {
-      setPaymentProcessing(false);
-      setProcessingMethod(null);
-    }
-  }, [createdReservation, breakdown, setPaymentProcessing, setPaymentError, t]);
-
-  const handleMercadoPago = useCallback(async () => {
-    if (!createdReservation || !breakdown) return;
-    setProcessingMethod('mercadopago');
-    setPaymentProcessing(true);
-    setPaymentError(null);
-
-    try {
-      const itemTitle = createdReservation.propertyId
-        ? `Nexo Real - Property Reservation`
-        : `Nexo Real - Tour Reservation`;
-
-      const result = await paymentService.createMercadoPagoPreference(
-        [
-          {
-            id: createdReservation.id,
-            title: itemTitle,
-            quantity: 1,
-            unit_price: breakdown.totalPrice,
-            currency_id: breakdown.currency,
-          },
-        ],
-        undefined,
-        createdReservation.id
-      );
-
-      paymentService.redirectToMercadoPago(result.initPoint);
-    } catch {
-      setPaymentError(t('reservation.paymentError'));
-    } finally {
-      setPaymentProcessing(false);
-      setProcessingMethod(null);
-    }
-  }, [createdReservation, breakdown, setPaymentProcessing, setPaymentError, t]);
-
-  return (
-    <div className="space-y-6 py-2">
-      {/* Header */}
-      <div className="text-center">
-        <h2 className="text-xl font-bold text-slate-800 mb-1">{t('reservation.selectPayment')}</h2>
-        <p className="text-slate-500 text-sm">{t('reservation.selectPaymentDesc')}</p>
-      </div>
-
-      {/* Total amount card */}
-      {breakdown && breakdown.totalPrice > 0 && <PriceBreakdownCard breakdown={breakdown} />}
-
-      {/* Payment error */}
-      {paymentError && (
-        <div className="flex items-start gap-2 bg-red-50 border border-red-200 rounded-lg p-3 text-sm text-red-600">
-          <AlertCircle className="w-4 h-4 mt-0.5 shrink-0" />
-          {paymentError}
-        </div>
-      )}
-
-      {/* Payment methods */}
-      <div className="relative space-y-3">
-        {/* PayPal */}
-        <Button
-          variant="outline"
-          onClick={handlePayPal}
-          disabled={isProcessingPayment}
-          className="w-full flex items-center gap-4 p-4 rounded-xl h-auto hover:border-emerald-300 hover:bg-emerald-50/30 group"
-        >
-          <div className="w-12 h-12 bg-blue-50 rounded-xl flex items-center justify-center shrink-0">
-            <span className="text-blue-600 font-bold text-lg">PP</span>
-          </div>
-          <div className="flex-1 text-left">
-            <span className="font-semibold text-slate-800 group-hover:text-emerald-700 transition-colors">
-              {t('reservation.payWithPaypal')}
-            </span>
-            <p className="text-xs text-slate-400 mt-0.5">{formatPrice(totalPrice, currency)}</p>
-          </div>
-          {processingMethod === 'paypal' ? (
-            <Loader2 className="w-5 h-5 text-emerald-500 animate-spin" />
-          ) : (
-            <ChevronRight className="w-5 h-5 text-slate-300 group-hover:text-emerald-500 transition-colors" />
-          )}
-        </Button>
-
-        {/* MercadoPago */}
-        <Button
-          variant="outline"
-          onClick={handleMercadoPago}
-          disabled={isProcessingPayment}
-          className="w-full flex items-center gap-4 p-4 rounded-xl h-auto hover:border-emerald-300 hover:bg-emerald-50/30 group"
-        >
-          <div className="w-12 h-12 bg-sky-50 rounded-xl flex items-center justify-center shrink-0">
-            <span className="text-sky-600 font-bold text-lg">MP</span>
-          </div>
-          <div className="flex-1 text-left">
-            <span className="font-semibold text-slate-800 group-hover:text-emerald-700 transition-colors">
-              {t('reservation.payWithMercadopago')}
-            </span>
-            <p className="text-xs text-slate-400 mt-0.5">{formatPrice(totalPrice, currency)}</p>
-          </div>
-          {processingMethod === 'mercadopago' ? (
-            <Loader2 className="w-5 h-5 text-emerald-500 animate-spin" />
-          ) : (
-            <ChevronRight className="w-5 h-5 text-slate-300 group-hover:text-emerald-500 transition-colors" />
-          )}
-        </Button>
-
-        {/* Wallet (hidden when crypto wallet feature is disabled) */}
-        {featureFlags.cryptoWallet && (
-          <Button
-            variant="outline"
-            onClick={onDone}
-            disabled={!hasEnoughBalance || isProcessingPayment}
-            className={cn(
-              'w-full flex items-center gap-4 p-4 rounded-xl h-auto group',
-              hasEnoughBalance
-                ? 'hover:border-emerald-300 hover:bg-emerald-50/30'
-                : 'border-slate-100 bg-slate-50 opacity-60 cursor-not-allowed'
-            )}
-          >
-            <div className="w-12 h-12 bg-emerald-50 rounded-xl flex items-center justify-center shrink-0">
-              <Wallet className="w-6 h-6 text-emerald-600" />
-            </div>
-            <div className="flex-1 text-left">
-              <span
-                className={cn(
-                  'font-semibold',
-                  hasEnoughBalance
-                    ? 'text-slate-800 group-hover:text-emerald-700 transition-colors'
-                    : 'text-slate-500'
-                )}
-              >
-                {t('reservation.payWithWallet')}
-              </span>
-              <p className="text-xs text-slate-400 mt-0.5">
-                {hasEnoughBalance
-                  ? t('reservation.walletBalance', {
-                      balance: formatPrice(walletBalance, walletCurrency),
-                    })
-                  : t('reservation.insufficientBalance')}
-              </p>
-            </div>
-            {hasEnoughBalance ? (
-              <ChevronRight className="w-5 h-5 text-slate-300 group-hover:text-emerald-500 transition-colors" />
-            ) : (
-              <AlertCircle className="w-5 h-5 text-slate-300" />
-            )}
-          </Button>
-        )}
-
-        {/* Payment processing overlay / Overlay de procesamiento de pago */}
-        {isProcessingPayment && (
-          <div
-            data-testid="payment-loading-overlay"
-            className="absolute inset-0 z-10 flex flex-col items-center justify-center rounded-xl bg-white/80 backdrop-blur-sm"
-          >
-            <Loader2 className="h-8 w-8 animate-spin text-emerald-500 mb-3" />
-            <p className="text-sm font-medium text-slate-700">{t('loading.processingPayment')}</p>
-          </div>
-        )}
-      </div>
-
-      {/* Security note */}
-      <div className="flex items-center gap-2 justify-center text-xs text-slate-400">
-        <Shield className="w-3.5 h-3.5" />
-        <span>
-          {t('reservation.secure')} — {t('reservation.secureDesc')}
-        </span>
-      </div>
-
-      {/* Pay later link */}
-      <div className="text-center">
-        <Button
-          variant="ghost"
-          onClick={onGoToReservations}
-          className="gap-2 text-sm text-slate-500 hover:text-emerald-600"
-        >
-          <Clock className="w-4 h-4" />
-          {t('reservation.payLater')}
-        </Button>
-      </div>
-    </div>
-  );
-}
 
 // ============================================
 // Main Page / Página principal
